requirements.txt
PyGithub
openai
requests 